fix comments for combined iceberg reader
